Save instance of WeatherStation
An instance of the WeatherStation class is now serialized to the
settings file and deserialized when it needs to be read. Some getter
methods were also added to the weatherStation class in order to properly
deserialize the weatherStation and obtain all necessary information.
Also fixed some tabbing differences in the file.